Add player stop and animate/rotate before using item.
Time to rotate is based on amount of degrees to rotate
(shorter time means less degrees).

Regular locomotion animation is used.
Player speed is synched.

If a special animation should be played,
projectile should be launched and player should be mobilized again
after animation is done, not when rotation is done.