Remove unused assets, and rename some.
Remove assets that are not used in the project anymore.
Assets that are still used are to be grouped better,
and badly named assets are renamed to increase
readability.

a branch 'refactor' is left in the master repository,
in case something breaks.
It contains lots of small commits that can be used to find where the breaking
change was introduced.