Removed autogenerated materials from Assets/Materials.
Whenever Unity finds files that is not in its cache,
it will import them.
Every time it does this, it will extract its materials,
create files for them, and link the models with the
new files.
This means that if a model is locally created,
and then the material is moved, then a subsequent
pull from another person will cause the models to
be reimported, and the old material files to be ignored.
This commit removes all such files.