Basic movement/jump/fall animation.
Player is in locomotion state until:
	1. user presses jump button
	2. player goes from grounded to not grounded
In case 1, player transitions to run jumping or idle jumping state based
on current speed. After 80% of the animation has played, it goes back to
locomotion.
In case 2, player transitions to falling. If player goes from not
grounded to grounded, it transitions back to locomotion.